+++ config/yaml.go
@@ -177,11 +177,6 @@ func (y yamlConfigProvider) Get(key string) Value {
 	return value
 }
 
-// Scope returns a scoped configuration provider
-func (y yamlConfigProvider) Scope(prefix string) Provider {
-	return NewScopedProvider(prefix, y)
-}
-
 func (y yamlConfigProvider) RegisterChangeCallback(key string, callback ChangeCallback) error {
 	// Yaml configuration don't receive callback events
 	return nil